renaming vue-filters to vue-utils
