+++ app/javascript/gobierto_plans/plugins/indicators.vue
@@ -25,11 +25,11 @@
 </template>
 
 <script>
-import { VueFiltersMixin } from "lib/shared";
+import { VueUtilsMixin } from "lib/shared";
 
 export default {
   name: "Indicators",
-  mixins: [VueFiltersMixin],
+  mixins: [VueUtilsMixin],
   props: {
     config: {
       type: Object,
