+++ app/javascript/lib/shared/index.js
@@ -15,7 +15,7 @@ import { HorizontalCarousel } from "./modules/horizontal-carousel.js";
 import { ImageLightbox } from "./modules/image-lightbox.js";
 import "./modules/separate-tabs.js";
 import { rangeSlider } from "./modules/range-slider.js";
-import { VueFiltersMixin } from "./modules/vue-filters.js";
+import { VueUtilsMixin } from "./modules/vue-utils.js";
 
 import accounting from "accounting";
 import { SETTINGS } from "./modules/accounting_settings.js";
@@ -33,5 +33,5 @@ export {
   readMore,
   HorizontalCarousel,
   ImageLightbox,
-  VueFiltersMixin
+  VueUtilsMixin
 };
